rs: spelling fixes

+++ reference/CodeExamples/java/src/com/db4odoc/concurrency/VersionConflictException.java
@@ -3,8 +3,8 @@ package com.db4odoc.concurrency;
 
 public class VersionConflictException extends RuntimeException{
 
-    public VersionConflictException(int committerVersion,
-                                    int currentVersion, Object object) {
+    public VersionConflictException(long committerVersion,
+                                    long currentVersion, Object object) {
         super("Committer-Version: "+committerVersion+
                 " Current Version: "+currentVersion +
                 " Object '"+object+"'"
